Politiques d'accès aux boutons 'edit' et 'delete'

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,12 @@ use Illuminate\Support\Facades\Auth;
 
 class QuestionsController extends Controller
 {
+
+    public function __construct() {
+        $this->middleware('auth')->except('index', 'show');                 // index() et show() n'ont pas besoin que l'usagé soit loggé
+    }
+
+
     /* Affiche toutes les questions */
     public function index() 
     {
@@ -56,6 +62,7 @@ class QuestionsController extends Controller
     /* Modifie une question */
     public function update(Request $request, Question $question)
     {
+        $this->authorize('delete', $question);                                // Utilise les restrictions définies dans App/policies/CustomersPolicy::delete()
         $data = request()->validate([
             'title' => 'required|max:255',
             'body' => 'required',
@@ -68,6 +75,7 @@ class QuestionsController extends Controller
 
     /* Supprime une question */
     public function destroy(Question $question) {
+        $this->authorize('delete', $question);                                // Utilise les restrictions définies dans App/policies/CustomersPolicy::delete()
         $question->delete();
         return redirect('questions')->with('success', 'Your question has been deleted');
     }
